requirements.txt
pytest>=5.4.1
pandas==1.0.3
ipython==7.8.0
azureml-sdk>=1.1.5
scikit-learn==0.20.3
# dash==0.21.1  
# dash-renderer==0.13.0  
# dash-html-components==0.11.0
# dash-core-components==0.23.0  
# plotly --upgrade
# Flask==1.1.1
azure-storage-blob==12.3.0
streamlit
seaborn